Add the Wikipedia politicians by country dataset
This data is obtained from Figshare on https://figshare.com/articles/dataset/Untitled_Item/5513449. The data contains most English-language Wikipedia articles within the category "Category: Politicians by nationality" and subcategories, along with the code used to generate that data. "country" column indicates the country of the politician, "page" column indicates the title of that Wikipedia page, and "rev_id" is the edit ID of the last edit to the page.